requirements.txt
matplotlib==3.2.2
numpy==1.19.2
pandas==1.1.2
pandas-datareader==0.9.0
PyYAML==5.3.1
requests==2.22.0
requests-cache==0.5.2
statsmodels==0.12.1
SQLAlchemy==1.3.19
